requirements.txt
hidapi>=0.14.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
